% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyty_copras.R
\name{.copras_defuzyfikuj_macierz}
\alias{.copras_defuzyfikuj_macierz}
\title{Wewnętrzna defuzyfikacja TFN dla COPRAS}
\usage{
.copras_defuzyfikuj_macierz(macierz_decyzyjna)
}
\description{
Zamienia macierz złożoną z trójek (l, m, u) na macierz ostrą
przez średnią arytmetyczną \eqn{(l + m + u) / 3}, zgodnie z procedurą
opisaną w dołączonej literaturze Fuzzy COPRAS.
}
\keyword{internal}
