% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyty_copras.R
\name{.copras_przygotuj_wagi}
\alias{.copras_przygotuj_wagi}
\title{Wewnętrzne przygotowanie wag dla COPRAS}
\usage{
.copras_przygotuj_wagi(
  macierz_decyzyjna,
  wagi,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\description{
Defuzyfikuje wagi do postaci ostrej i normalizuje je do sumy 1.
}
\keyword{internal}
