% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_solver.R
\name{.dodaj_ograniczenie}
\alias{.dodaj_ograniczenie}
\title{Pomocnik budowania ograniczeń}
\usage{
.dodaj_ograniczenie(ograniczenia, nowe_ograniczenie)
}
\description{
Pomocnik budowania ograniczeń
}
\keyword{internal}
