% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{.oblicz_ranking_dominacji}
\alias{.oblicz_ranking_dominacji}
\title{Teoria Dominacji dla Rankingu}
\usage{
.oblicz_ranking_dominacji(r1, r2)
}
\arguments{
\item{r1}{Wektor numeryczny rang metody 1.}

\item{r2}{Wektor numeryczny rang metody 2.}
}
\value{
Wektor numeryczny z finalnym rankingiem.
}
\description{
Funkcja pomocnicza. Wyznacza ranking konsensusu na podstawie reguły większości.
Iteracyjnie sprawdza, która alternatywa najczęściej wygrywa na danej pozycji.
}
\keyword{internal}
