% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{.pobierz_finalne_wagi}
\alias{.pobierz_finalne_wagi}
\title{Wewnętrzny procesor wag}
\usage{
.pobierz_finalne_wagi(
  macierz,
  wagi,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\description{
Decyduje, skąd wziąć wagi (Ręczne vs BWM).
}
\keyword{internal}
