% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.rozmyj_wektor}
\alias{.rozmyj_wektor}
\title{Wewnętrzna funkcja rozmywająca (Fuzzifier)}
\usage{
.rozmyj_wektor(wektor)
}
\description{
Zamienia liczbę rzeczywistą (Crisp) na Trójkątną Liczbę Rozmytą (TFN).
TFN to trójka (l, m, u), gdzie m = x, l = x-1, u = x+1.
}
\keyword{internal}
