% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.skaluj_do_saaty}
\alias{.skaluj_do_saaty}
\title{Wewnętrzny Skaler Saaty'ego}
\usage{
.skaluj_do_saaty(wektor)
}
\description{
Przekształca dowolną skalę na skalę Saaty'ego 1-9.
}
\keyword{internal}
