% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_solver.R
\name{oblicz_wagi_bwm}
\alias{oblicz_wagi_bwm}
\title{Obliczanie wag metodą BWM}
\usage{
oblicz_wagi_bwm(nazwy_kryteriow, najlepsze_do_innych, inne_do_najgorszego)
}
\arguments{
\item{nazwy_kryteriow}{Wektor znakowy z nazwami kryteriów.}

\item{najlepsze_do_innych}{Wektor numeryczny (1-9). Preferencja Najlepszego kryterium nad innymi.}

\item{inne_do_najgorszego}{Wektor numeryczny (1-9). Preferencja innych kryteriów nad Najgorszym.}
}
\value{
Lista zawierająca: `wagi_kryteriow`, `wskaznik_spojnosci` (CR) oraz wartość `ksi`.
}
\description{
Wyznacza optymalne wagi kryteriów metodą Best-Worst (BWM) przy użyciu
programowania liniowego. Minimalizuje wskaźnik niespójności (ksi).
}
