% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{oblicz_wagi_entropii}
\alias{oblicz_wagi_entropii}
\title{Obliczanie wag metodą Entropii Shannona}
\usage{
oblicz_wagi_entropii(macierz_decyzyjna)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz (wynik funkcji `przygotuj_dane_mcda`).}
}
\value{
Wektor numeryczny wag sumujący się do 1.
}
\description{
Wyznacza obiektywne wagi kryteriów na podstawie danych,
mierząc stopień rozproszenia wartości. Im większa zmienność, tym wyższa waga.
}
