% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_copras_promo_wynik}
\alias{plot.rozmyty_copras_promo_wynik}
\title{Mapa Strategiczna COPRAS}
\usage{
\method{plot}{rozmyty_copras_promo_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_copras_promo_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Wizualizacja wyników metody COPRAS w układzie dwuwymiarowym.
Oś X: Suma korzyści (S+). Oś Y: Suma kosztów (S-).
Wielkość bąbla: Stopień użyteczności (Ui).
}
