% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_topsis_promo_wynik}
\alias{plot.rozmyty_topsis_promo_wynik}
\title{Mapa Strategiczna TOPSIS}
\usage{
\method{plot}{rozmyty_topsis_promo_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_topsis_promo_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Wizualizacja wyników metody TOPSIS w układzie dwuwymiarowym.
}
