% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{przygotuj_dane_mcda}
\alias{przygotuj_dane_mcda}
\title{Przygotowanie Danych do Rozmytej Analizy MCDA}
\usage{
przygotuj_dane_mcda(
  dane,
  skladnia,
  kolumna_alternatyw = NULL,
  funkcja_agregacji = mean
)
}
\arguments{
\item{dane}{Ramka danych (data frame) zawierająca surowe zmienne.}

\item{skladnia}{Ciąg znaków definiujący kryteria (np. "Koszt =~ k1 + k2").}

\item{kolumna_alternatyw}{Nazwa kolumny identyfikującej alternatywy.
Jeśli NULL, każdy wiersz traktowany jest jako osobna alternatywa.}

\item{funkcja_agregacji}{Funkcja używana do scalania opinii ekspertów (domyślnie: mean).}
}
\value{
Macierz o wymiarach ($m \times 3n$), gdzie m to liczba alternatyw.
}
\description{
Funkcja przekształca surowe dane ankietowe w rozmytą macierz decyzyjną.
Oblicza wyniki zmiennych kompozytowych na podstawie składni, skaluje je do przedziału 1-9,
agreguje odpowiedzi ekspertów (jeśli dotyczy) i dokonuje rozmycia (fuzzification).
}
