% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyty_copras.R
\name{rozmyty_copras_promo}
\alias{rozmyty_copras_promo}
\title{Rozmyta Metoda COPRAS (Complex Proportional Assessment)}
\usage{
rozmyty_copras_promo(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz decyzyjna (wynik funkcji przygotuj_dane_mcda).}

\item{typy_kryteriow}{Wektor tekstowy określający charakter kryteriów ("min" lub "max").}

\item{wagi}{Opcjonalny wektor wag. Jeśli brak, zostaną pobrane z atrybutów lub BWM.}

\item{bwm_kryteria}{(Opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_copras_promo_wynik`.
}
\description{
Funkcja realizuje algorytm Fuzzy COPRAS do oceny wielokryterialnej.
Zgodnie z procedurą opisaną w dołączonej literaturze, macierz rozmyta jest
najpierw defuzyfikowana do postaci ostrej, a następnie poddawana klasycznym
krokom COPRAS: normalizacji przez sumy kolumn, ważeniu oraz wyznaczeniu
wskaźników \eqn{P_i}, \eqn{R_i}, \eqn{Q_i} i \eqn{U_i}.
}
