% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{rozmyty_meta_ranking}
\alias{rozmyty_meta_ranking}
\title{Rozmyty Meta-Ranking}
\usage{
rozmyty_meta_ranking(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL,
  lambda = 0.5,
  v = 0.5
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz danych.}

\item{typy_kryteriow}{Wektor typów ("min", "max").}

\item{wagi}{(Opcjonalnie) Wagi kryteriów.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor BWM Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor BWM Others-to-Worst.}

\item{lambda}{Parametr dla WASPAS (domyślnie 0.5).}

\item{v}{Parametr dla VIKOR (domyślnie 0.5).}
}
\value{
Lista zawierająca ramkę danych z porównaniem rankingów oraz macierz korelacji.
}
\description{
Agreguje wyniki z metod Fuzzy VIKOR, TOPSIS i WASPAS, aby stworzyć
jeden, robustny ranking konsensusu.
}
