% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{rozmyty_vikor_promo}
\alias{rozmyty_vikor_promo}
\title{Rozmyta Metoda VIKOR}
\usage{
rozmyty_vikor_promo(
  macierz_decyzyjna,
  typy_kryteriow,
  v = 0.5,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz o wymiarach \eqn{m \times 3n}.}

\item{typy_kryteriow}{Wektor znakowy ("max" dla zysku, "min" dla kosztu).}

\item{v}{Waga strategii "większości kryteriów" (domyślnie 0.5).}

\item{wagi}{(Opcjonalnie) Wektor wag.}

\item{bwm_kryteria}{(Opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_vikor_promo_wynik`.
}
\description{
Metoda kompromisowa VIKOR. Oblicza wskaźniki S (użyteczność grupy),
R (indywidualny żal) oraz Q (indeks kompromisu).
}
