% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{.oblicz_ranking_dominacji}
\alias{.oblicz_ranking_dominacji}
\title{Teoria Dominacji dla Rankingu}
\usage{
.oblicz_ranking_dominacji(macierz_rang)
}
\arguments{
\item{macierz_rang}{Macierz (wiersze = alternatywy, kolumny = metody).}
}
\value{
Wektor numeryczny z finalnym rankingiem.
}
\description{
Funkcja pomocnicza. Wyznacza ranking konsensusu na podstawie reguły większości.
Iteracyjnie sprawdza, która alternatywa najczęściej wygrywa na danej pozycji.
}
\keyword{internal}
