% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{rozmyty_copras_promo}
\alias{rozmyty_copras_promo}
\title{Rozmyta Metoda COPRAS (Complex Proportional Assessment)}
\usage{
rozmyty_copras_promo(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria,
  bwm_najlepsze,
  bwm_najgorsze
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz decyzyjna (wynik funkcji przygotuj_dane_mcda).}

\item{typy_kryteriow}{Wektor tekstowy określający charakter kryteriów ("min" lub "max").}

\item{wagi}{Opcjonalny wektor wag. Jeśli brak, zostaną pobrane z atrybutów lub BWM.}

\item{bwm_kryteria}{(Opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(Opcjonalnie) Wektor Best-to-Others.}

\item{bwm_najgorsze}{(Opcjonalnie) Wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_copras_promo_wynik`.
}
\description{
Funkcja realizuje algorytm Fuzzy COPRAS do oceny wielokryterialnej.
Metoda ta wyznacza ranking alternatyw na podstawie ich względnego priorytetu (Qi)
oraz stopnia użyteczności (Ui).
}
